python/packages/botas/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "botas"
version = "0.1.0"
description = "Lightweight library for building Microsoft Bot Framework bots - Python"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "httpx>=0.27",
    "PyJWT[crypto]>=2.8",
    "msal>=1.28",
    "pydantic>=2.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src/botas"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
